fix bug starting up API and bug with dependency injection for Search
