+++ LBHTenancyAPI/Controllers/V2/SearchController.cs
@@ -11,8 +11,8 @@ namespace LBHTenancyAPI.Controllers.V2
     /// <summary>
     /// Search Controller V2 to search for Tenants in a multi faceted way
     /// </summary>
-    [ApiVersion("2")]
-    [Route("api/v2/tenancies/search/")]
+    [ApiVersion("1.1")]
+    [Route("api/v{version:apiVersion}/tenancies/search/", Name = "SearchV2")]
     [Produces("application/json")]
     [ProducesResponseType(typeof(APIResponse<object>), 400)]
     [ProducesResponseType(typeof(APIResponse<object>), 500)]
@@ -41,8 +41,7 @@ namespace LBHTenancyAPI.Controllers.V2
         /// </summary>
         /// <param name="request"></param>
         /// <returns></returns>
-        [HttpGet]
-        [MapToApiVersion("2")]
+        [HttpGet, MapToApiVersion("1.1")]
         [ProducesResponseType(typeof(APIResponse<SearchTenancyResponse>), 200)]
         public async Task<IActionResult> Get([FromQuery]SearchTenancyRequest request)
         {
