+++ LBHTenancyAPI/Startup.cs
@@ -1,13 +1,20 @@
 using System;
 using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Reflection;
+using LBHTenancyAPI.Extensions.Versioning;
 using LBHTenancyAPI.Infrastructure.V1.Services;
 using LBHTenancyAPI.Middleware;
 using LBHTenancyAPI.Settings;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ApiExplorer;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Swashbuckle.AspNetCore.Swagger;
+using Swashbuckle.AspNetCore.SwaggerGen;
 
 namespace LBHTenancyAPI
 {
@@ -29,7 +36,7 @@ namespace LBHTenancyAPI
         }
 
         public IConfiguration Configuration { get; }
-        private List<ApiVersionDescription> _apiVersions { get; set; }
+        private static List<ApiVersionDescription> _apiVersions { get; set; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
@@ -55,7 +62,53 @@ namespace LBHTenancyAPI
             services.ConfigureApiVersioning();
 
             //Set [ApiVersion("x")] on Controllers to automatically add them to swagger docs
-            services.ConfigureSwaggerGen();
+            services.AddSingleton<IApiVersionDescriptionProvider, DefaultApiVersionDescriptionProvider>();
+
+            //add swagger gen to generate the swagger.json file - delayed execution
+            // Automatically Generates Swagger docs with XML comments based on the [ApiVersion("x")] on a controller
+            // and assigns that 
+            services.AddSwaggerGen(c =>
+            {
+                c.AddSecurityDefinition("Token",
+                    new ApiKeyScheme
+                    {
+                        In = "header",
+                        Description = "Your Hackney API Key",
+                        Name = "X-Api-Key",
+                        Type = "apiKey"
+                    });
+                c.AddSecurityRequirement(new Dictionary<string, IEnumerable<string>>
+                {
+                    {"Token", Enumerable.Empty<string>()}
+                });
+
+                //Looks at the APIVersionAttribute [ApiVersion("x")] on controllers and decides whether or not
+                //to include it in that version of the swagger document
+                //Controllers must have this [ApiVersion("x")] to be included in swagger documentation!!
+                c.DocInclusionPredicate((docName, apiDesc) =>
+                {
+                    var versions = apiDesc.ControllerAttributes()
+                        .OfType<ApiVersionAttribute>()
+                        .SelectMany(attr => attr.Versions).ToList();
+
+                    var any = versions.Any(v => $"{v.GetFormattedApiVersion()}" == docName);
+                    return any;
+                });
+
+                //Get every ApiVersion attribute specified and create swagger docs for them
+                foreach (var apiVersion in _apiVersions)
+                {
+                    var version = $"v{apiVersion.ApiVersion.ToString()}";
+                    c.SwaggerDoc(version, new Info { Title = $"TenancyAPI {version}", Version = version });
+                }
+
+                c.CustomSchemaIds(x => x.FullName);
+                // Set the comments path for the Swagger JSON and UI.
+                var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
+                var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
+                if (File.Exists(xmlPath))
+                    c.IncludeXmlComments(xmlPath);
+            });
 
 
             services.ConfigureLogging(Configuration, settings);
@@ -74,7 +127,7 @@ namespace LBHTenancyAPI
             //Register exception handling middleware early so exceptions are handled and formatted
             app.UseMiddleware<CustomExceptionHandlerMiddleware>();
 
-            app.ConfigureSwaggerUI(_apiVersions);
+            _apiVersions = app.ConfigureSwaggerUI(_apiVersions);
 
             //required for swagger to work
             app.UseMvc(routes =>
